basic component complete, more functionality to come but pushing to get displayed
